TINKERPOP-2769 Add response messages to handlers when cancelling tasks.
It is currently possible for requests to get cancelled for exceeding the
evaluationTimeout even before its associated response task starts running.
This can cause clients to hang as they wait for a response from a cancelled
task. Checks are added in this commit to tell if the task has started to run
and if they haven't then they will send a timeout response. If the task has
started running then nothing is done to prevent behavior change.