Dockerize gremlin-console python-based integration tests.
Python 3.8 is used for the image because the version of pytest used
isn't compatible with later versions of Python. Building our own Docker
image also allows for greater control over the version of Python and
Java used.

The location of the python test output was changed in order to fit
better with the bind mounted volume for Docker.