Update docker/build.sh to work with GLV docker-compose changes.
The non-Java GLVs have changed to use docker-compose for integration
testing which creates a Docker in Docker issue when run through the
build.sh script. To prevent this issue, the GLVs are split away and run
individually using their docker-compose files.

To increase the usability of the script for developers who only work on
a portion of TinkerPop, options were added to include only a subset of
the modules to the build.