Updated base images for gremlin-server and gremlin-console to support arm64
The previously used adopt openjdk11 alpine linux image does not support arm64.
In fact none of the major openjdk distrobutions maintain a jdk11 alpine image for arm64.

Changed base image to alpine and installed oracle openjdk 11. Oracle openjdk 11 was chosen as it is
already maintained in the APK community repository.