Downgrade jansi to v1.17.1 to fix colored output in console
Jansi was declared as an explicit dependency of gremlin-console in order to resolve
build issues on arm macs. Initially v2.1.0 was added but this version is missing methods
which the console was using for colorizing output. To sidestep this a try catch was added to Colorizer.groovy
to catch this exception and instead return uncolored text.

This commit changes the jansi version to 1.17.1 which still includes the needed methods for colorizing output.
Adding Jansi 1.17.1 as a dependency resolves the build issues on arm macs without breaking colorizing or requiring
any changes to source code. For this reason the previous changes to Colorizer.groovy were reverted.