Removed unneeded information from docs
Removed information describing the neccessary changes for M1 compatibility
as these steps will not be neccessary once the changes are merged in.

Moved ARM mac docker configuration tips to the testing with docker section where it is more relevant
and because there is no longer enough M1 specific information to justify its own section in the docs.