Created gremlin-util module and refactored gremlin-driver
The new gremlin-util module is intended to house any code which is to be shared between gremlin-driver and gremlin-server.

The motivation behind this change is to eventually create a robust test server to be used by all GLV's without requiring gremlin-driver to be a dependency of the GLVs.
A side motivation is to remove the existing dependency of gremlin-server on gremlin-driver or at least reduce the scope of the dependency to test.

At this stage the only changes are moving all of the shared classes to the new module but leaving them in their old tinkerpop.gremlin.driver packages.
While it is an odd choice to have the package name not correlate to the module, the reasoning behind it is to maintain compatibility with 3.5 while assessing the viability and value of these changes.
The packages should be changed to correspond with the new module in a future version where such a breaking change is permittable.

This refactor also reduces gremlin-server's dependency on gremlin-driver to a test scope only.
The only issue with reducing this dependency is one javadoc @Link to gremlin-driver/Client. It does not seem worthwhile to maintain a dependency between server and driver for one javadoc comment