Dockerized gremlin-socket-server and decouple from gremlin-driver
Builds SimpleSocketServer into a docker image and reconfigures the java driver to run this container during integration tests
Moves hardcoded constants for the server port and custom request id's into a config yaml to decouple gremlin-socket-server from any drivers which use it for tests.
These changes bring gremlin-socket-server into a state which can be utilized by all of the GLV's as opposed to only the java driver.