TINKERPOP-2946 Detect ordering issues in Gherkin tests
This commit addresses part 1. of TINKERPOP-2946 by introducing
TinkerShuffleGraph and running feature tests against an embedded instance
of this graph.

TinkerShuffleGraph is derived from TinkerGraph and intended for test purposes only.
It's only differentiation from TinkerGraph is that it will returned shuffled results
when getting any of the following:

vertices from the graph,
edges from the graph,
vertices from a vertex,
edges from a vertex,
vertices from an edge,
properties from a vertex,
properties from an edge, and
properties from a VertexProperty